requirements.txt
# langchain==0.0.154
langchain==0.0.316
langchain-community==0.0.14
# langchain-openai==0.0.4
langchain-core>=0.1.14,<0.2
openai==0.28
# openai==1.7
httpx==0.23
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
altair<5
tiktoken
joblib
